+++ database/migrations/2022_08_06_064838_add_column_to_gostaresh_academic_major_educational_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('gostaresh_academic_major_educational', function (Blueprint $table) {
-            $table->unsignedBigInteger('medical_sciences_count')->default(0);
+            $table->float('medical_sciences_count')->default(0);
             $table->float('medical_sciences_percent')->nullable();
         });
     }
